requirements.txt
asgiref==3.6.0
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
Django==4.1.7
djangorestframework==3.14.0
executing==1.2.0
Flask==2.2.3
fonttools==4.39.2
idna==3.4
importlib-metadata==6.3.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
numpy==1.24.2
packaging==23.0
pickleshare==0.7.5
Pillow==9.4.0
pure-eval==0.2.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
pywin32==305
pyzmq==25.0.0
requests==2.28.2
scikit-learn==1.2.1
scipy==1.10.1
six==1.16.0
sqlparse==0.4.3
threadpoolctl==3.1.0
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.14
wcwidth==0.2.6
Werkzeug==2.2.3
zipp==3.15.0
